First Demo Version! 5 fully playable level

+++ Assets/Scripts/Entities/Bombs/Bomb_Cross.cs
@@ -9,8 +9,8 @@ public class Bomb_Cross : Bomb
     {
         AffectedArea.Clear();
 
-        float x = parentTile.localPosition.x;
-        float y = parentTile.localPosition.y;
+        int x = (int)parentTile.localPosition.x;
+        int y = (int)parentTile.localPosition.y;
 
 
         //print($"{x},{y}");
