+++ Assets/Scripts/Entities/Bombs/Bomb_X.cs
@@ -8,11 +8,11 @@ public class Bomb_X : Bomb
     {
         AffectedArea.Clear();
 
-        float x = parentTile.localPosition.x;
-        float y = parentTile.localPosition.y;
+        int x = (int)parentTile.localPosition.x;
+        int y = (int)parentTile.localPosition.y;
 
 
-       // print($"{x},{y}");
+        // print($"{x},{y}");
         foreach (Transform tile in GridManager.Instance.Tiles)
         {
             for (int i = 0; i < _range; i++)
@@ -31,6 +31,8 @@ public class Bomb_X : Bomb
         {
             tile.gameObject.GetComponent<AffectedAreaOutline>().enabled = true;
         }
+
+
     }
 
 }
